Move FromDouble method to a separate file: Fraction.ConvertFromDouble.cs (was a suggestion from @lipchev in a previous PR - which simplifies code reviews).
